// frontend/app/src/app/models/ProjekatZaIzmenu.ts
export class ProjektiZaIzmenu{
    idProjekta: number = 0;
    nazivProjekta: string = "";
    datumProjekta: string = "" ;
    akronim: string = "";
    apstraktSrp: string = "";
    apstraktEng: string = "";
    ukupanBudzet: number = 0;
    snimanjeProjekta: string = "";
    podprogram: string = "";
    idRukovodioca: number = 0;
    nazivInstSrp: string = "";
    projekatInst1: number = 0;
    projekatInst2: number = 0;
    projekatInst3: number = 0;
    projekatInst4: number = 0;
    projekatInst5: number = 0;
    projekatOblast1: number = 0;
    projekatOblast2: number = 0;
    projekatOblast3: number = 0;
    projekatOblast4: number = 0;
    projekatOblast5: number = 0;
    status: string = ""; 
    
}